Revert "Revert "intel/package.use: app-text/mupdf +X""
As discussed on the ML.

It adds a menu item for the GUI tool, is pulled in by default (because
of net-print/cups-filters) but is useful to some.

If it should not be visible to users by default, it could be hidden from
the DE or forked to stop installing the menu file.

This reverts commit 9e4b41ed5ca9677e5095c2c2d7b755e061a779ec.